Channel list groups are minimal in height with one greedy.
This updates the channel list groups from being equi-sized to
determining the vertical height to show for each group.  The general
approach here is the expectation that most groups are relatively short
except the main public channels, so the former are sized to only use
the needed height and the public channels greedily use the rest of the
vertical space.  If the total amount of vertical space (dynamically
determined) is too small to allow fair full rendering of the other
groups, all groups are truncated.